+++ src/asf_tools/hydrosar/flood_map.py
@@ -15,7 +15,7 @@ import tempfile
 import warnings
 from pathlib import Path
 from shutil import make_archive
-from typing import Callable, Literal, Tuple, Union
+from typing import Callable, Literal, Optional, Tuple, Union
 
 import numpy as np
 from osgeo import gdal
@@ -23,20 +23,20 @@ from scipy import ndimage, optimize, stats
 from tqdm import tqdm
 
 from asf_tools.aws import get_path_to_s3_file, upload_file_to_s3
-from asf_tools.composite import get_epsg_code, write_cog
-from asf_tools.raster import read_as_masked_array
-
+from asf_tools.raster import read_as_masked_array, write_cog
+from asf_tools.util import get_coordinates, get_epsg_code
 
 log = logging.getLogger(__name__)
 
 
-def get_coordinates(info: dict) -> Tuple[int, int, int, int]:
-    west, south = info['cornerCoordinates']['lowerLeft']
-    east, north = info['cornerCoordinates']['upperRight']
-    return west, south, east, north
+def get_pw_threshold(water_array: np.array) -> float:
+    hist, bin_edges = np.histogram(water_array, density=True, bins=100)
+    reverse_cdf = np.cumsum(np.flipud(hist)) * (bin_edges[1] - bin_edges[0])
+    ths_orig = np.flipud(bin_edges)[np.searchsorted(np.array(reverse_cdf), 0.95)]
+    return round(ths_orig) + 1
 
 
-def get_waterbody(input_info: dict, threshold: float = 30.) -> np.array:
+def get_waterbody(input_info: dict, threshold: Optional[float] = None) -> np.array:
     epsg = get_epsg_code(input_info)
 
     west, south, east, north = get_coordinates(input_info)
@@ -51,16 +51,29 @@ def get_waterbody(input_info: dict, threshold: float = 30.) -> np.array:
                   width=width, height=height, resampleAlg='nearest', format='GTiff')
         water_array = gdal.Open(water_extent_file.name, gdal.GA_ReadOnly).ReadAsArray()
 
+    if threshold is None:
+        threshold = get_pw_threshold(water_array)
+
     return water_array > threshold
 
 
-def iterative(hand: np.array, extent: np.array, water_levels: np.array = range(15)):
-    def _goal_ts(w):
+def iterative(hand: np.array, extent: np.array, water_levels: np.array = np.arange(15),
+              minimization_metric: str = 'ts'):
+    def get_confusion_matrix(w):
         iterative_flood_extent = hand < w  # w=water level
         tp = np.nansum(np.logical_and(iterative_flood_extent == 1, extent == 1))  # true positive
+        tn = np.nansum(np.logical_and(iterative_flood_extent == 0, extent == 0))  # true negative
         fp = np.nansum(np.logical_and(iterative_flood_extent == 1, extent == 0))  # False positive
         fn = np.nansum(np.logical_and(iterative_flood_extent == 0, extent == 1))  # False negative
-        return 1 - tp / (tp + fp + fn)  # threat score #we will minimize goal func, hence 1-threat_score.
+        return tp, tn, fp, fn
+
+    def _goal_ts(w):
+        tp, _, fp, fn = get_confusion_matrix(w)
+        return 1 - tp / (tp + fp + fn)  # threat score -- we will minimize goal func, hence `1 - threat_score`.
+
+    def _goal_fmi(w):
+        tp, _, fp, fn = get_confusion_matrix(w)
+        return 1 - np.sqrt((tp/(tp+fp))*(tp/(tp+fn)))
 
     class MyBounds(object):
         def __init__(self, xmax=max(water_levels), xmin=min(water_levels)):
@@ -74,15 +87,17 @@ def iterative(hand: np.array, extent: np.array, water_levels: np.array = range(1
             return tmax and tmin
 
     bounds = MyBounds()
-    temp_wl = np.zeros(max(water_levels))
-    for i in range(1, max(water_levels)):
-        opt_res = optimize.basinhopping(_goal_ts, i, niter=10000, niter_success=100, accept_test=bounds)
-        if opt_res.message[0] == 'success condition satisfied' \
-                or opt_res.message[0] == 'requested number of basinhopping iterations completed successfully':
-            temp_wl[i] = opt_res.x[0]
-        else:
-            temp_wl[i] = np.inf  # set as inf to mark unstable solution
-    return np.nanmean(temp_wl)
+    MINIMIZATION_FUNCTION = {'fmi': _goal_fmi, 'ts': _goal_ts}
+    with warnings.catch_warnings():
+        warnings.filterwarnings('ignore', category=RuntimeWarning)
+        opt_res = optimize.basinhopping(MINIMIZATION_FUNCTION[minimization_metric], np.mean(water_levels),
+                                        niter=10000, niter_success=100, accept_test=bounds, stepsize=3)
+
+    if opt_res.message[0] == 'success condition satisfied' \
+            or opt_res.message[0] == 'requested number of basinhopping iterations completed successfully':
+        return opt_res.x[0]
+    else:
+        return np.inf  # set as inf to mark unstable solution
 
 
 def logstat(data: np.ndarray, func: Callable = np.nanstd) -> Union[np.ndarray, float]:
@@ -101,22 +116,28 @@ def logstat(data: np.ndarray, func: Callable = np.nanstd) -> Union[np.ndarray, f
     return np.exp(st)
 
 
-def estimate_flood_depth(label, hand, flood_labels, estimator='iterative', water_level_sigma=3.,
-                         iterative_bounds=(0, 15)):
+def estimate_flood_depth(label: int, hand: np.ndarray, flood_labels: np.ndarray, estimator: str = 'iterative',
+                         water_level_sigma: float = 3., iterative_bounds: Tuple[int, int] = (0, 15),
+                         iterative_min_size: int = 0, minimization_metric: str = 'ts') -> float:
     with warnings.catch_warnings():
         warnings.filterwarnings('ignore', r'Mean of empty slice')
 
         if estimator.lower() == "iterative":
-            return iterative(hand, flood_labels == label, water_levels=iterative_bounds)
+            if (flood_labels == label).sum() < iterative_min_size:
+                return np.nan
+
+            water_levels = np.arange(*iterative_bounds)
+            return iterative(hand, flood_labels == label,
+                             water_levels=water_levels, minimization_metric=minimization_metric)
+
+        if estimator.lower() == "nmad":
+            hand_mean = np.nanmean(hand[flood_labels == label])
+            hand_std = stats.median_abs_deviation(hand[flood_labels == label], scale='normal', nan_policy='omit')
 
         if estimator.lower() == "numpy":
             hand_mean = np.nanmean(hand[flood_labels == label])
             hand_std = np.nanstd(hand[flood_labels == label])
 
-        elif estimator.lower() == "nmad":
-            hand_mean = np.nanmean(hand[flood_labels == label])
-            hand_std = stats.median_abs_deviation(hand[flood_labels == label], scale='normal',
-                                                  nan_policy='omit')
         elif estimator.lower() == "logstat":
             hand_mean = logstat(hand[flood_labels == label], func=np.nanmean)
             hand_std = logstat(hand[flood_labels == label])
@@ -127,12 +148,15 @@ def estimate_flood_depth(label, hand, flood_labels, estimator='iterative', water
     return hand_mean + water_level_sigma * hand_std
 
 
-def make_flood_map(out_raster: Union[str, Path],  vv_raster: Union[str, Path],
+def make_flood_map(out_raster: Union[str, Path], vv_raster: Union[str, Path],
                    water_raster: Union[str, Path], hand_raster: Union[str, Path],
                    estimator: str = 'iterative',
                    water_level_sigma: float = 3.,
-                   known_water_threshold: float = 30.,
-                   iterative_bounds: Tuple[int, int] = (0, 15)):
+                   known_water_threshold: Optional[float] = None,
+                   iterative_bounds: Tuple[int, int] = (0, 15),
+                   iterative_min_size: int = 0,
+                   minimization_metric: str = 'ts',
+                   ):
     """Create a flood depth map from a surface water extent map.
 
     WARNING: This functionality is still under active development and the products
@@ -164,8 +188,14 @@ def make_flood_map(out_raster: Union[str, Path],  vv_raster: Union[str, Path],
         hand_raster: Height Above Nearest Drainage (HAND) GeoTIFF aligned to the surface water extent raster
         estimator: Estimation approach for determining flood depth
         water_level_sigma: Max water height used in logstat, nmad, and numpy estimations
-        known_water_threshold: Threshold for extracting the known water area in percent
-        iterative_bounds: Bounds on basin-hopping algorithm used in iterative estimation
+        known_water_threshold: Threshold for extracting the known water area in percent.
+            If `None`, the threshold is calculated.
+        iterative_bounds: Minimum and maximum bound on the flood depths calculated by the basin-hopping algorithm
+            used in the iterative estimator
+        iterative_min_size: Minimum size of a connected waterbody in pixels for calculating flood depths with the
+            iterative estimator. Waterbodies smaller than this wll be skipped.
+        minimization_metric: Evaluation method to minimize when using the iterative estimator.
+            Options include a Fowlkes-Mallows index (fmi) or a threat score (ts).
 
     References:
         Jean-Francios Pekel, Andrew Cottam, Noel Gorelik, Alan S. Belward. 2016. <https://doi:10.1038/nature20584>
@@ -174,7 +204,6 @@ def make_flood_map(out_raster: Union[str, Path],  vv_raster: Union[str, Path],
     info = gdal.Info(str(water_raster), format='json')
     epsg = get_epsg_code(info)
     geotransform = info['geoTransform']
-
     hand_array = gdal.Open(str(hand_raster), gdal.GA_ReadOnly).ReadAsArray()
 
     log.info('Fetching perennial flood data.')
@@ -193,7 +222,9 @@ def make_flood_map(out_raster: Union[str, Path],  vv_raster: Union[str, Path],
 
     labeled_flood_mask, num_labels = ndimage.label(flood_mask)
     object_slices = ndimage.find_objects(labeled_flood_mask)
-    log.info(f'Detected {num_labels} water bodies...')
+    log.info(f'Detected {num_labels} waterbodies...')
+    if estimator.lower() == 'iterative':
+        log.info(f'Skipping waterbodies less than {iterative_min_size} pixels.')
 
     flood_depth = np.zeros(flood_mask.shape)
 
@@ -205,8 +236,11 @@ def make_flood_map(out_raster: Union[str, Path],  vv_raster: Union[str, Path],
         flood_window = labeled_flood_mask[min0:max0, min1:max1]
         hand_window = hand_array[min0:max0, min1:max1]
 
-        water_height = estimate_flood_depth(ll, hand_window, flood_window, estimator=estimator,
-                                            water_level_sigma=water_level_sigma, iterative_bounds=iterative_bounds)
+        water_height = estimate_flood_depth(
+            ll, hand_window, flood_window, estimator=estimator, water_level_sigma=water_level_sigma,
+            iterative_bounds=iterative_bounds, minimization_metric=minimization_metric,
+            iterative_min_size=iterative_min_size,
+        )
 
         flood_depth_window = flood_depth[min0:max0, min1:max1]
         flood_depth_window[flood_window == ll] = water_height - hand_window[flood_window == ll]
@@ -232,6 +266,18 @@ def make_flood_map(out_raster: Union[str, Path],  vv_raster: Union[str, Path],
               epsg_code=epsg, dtype=gdal.GDT_Float64, nodata_value=nodata)
 
 
+def optional_str(value: str) -> Optional[str]:
+    if value.lower() == 'none':
+        return None
+    return value
+
+
+def optional_float(value: str) -> Optional[float]:
+    if value.lower() == 'none':
+        return None
+    return float(value)
+
+
 def _get_cli(interface: Literal['hyp3', 'main']) -> argparse.ArgumentParser:
     parser = argparse.ArgumentParser(
         description=__doc__,
@@ -239,12 +285,14 @@ def _get_cli(interface: Literal['hyp3', 'main']) -> argparse.ArgumentParser:
     )
 
     available_estimators = ['iterative', 'logstat', 'nmad', 'numpy']
+    estimator_help = 'Flood depth estimation approach.'
     if interface == 'hyp3':
         parser.add_argument('--bucket')
         parser.add_argument('--bucket-prefix', default='')
         parser.add_argument('--wm-raster',
                             help='Water map GeoTIFF raster, with suffix `_WM.tif`.')
-        available_estimators.append('None')
+        available_estimators.append(None)
+        estimator_help += ' If `None`, flood depth will not be calculated.'
     elif interface == 'main':
         parser.add_argument('out_raster',
                             help='File to which flood depth map will be saved.')
@@ -258,19 +306,29 @@ def _get_cli(interface: Literal['hyp3', 'main']) -> argparse.ArgumentParser:
     else:
         raise NotImplementedError(f'Unknown interface: {interface}')
 
-    parser.add_argument('--estimator', type=str, default='iterative', choices=available_estimators,
-                        help='Flood depth estimation approach.')
+    parser.add_argument('--estimator', type=optional_str, default='iterative', choices=available_estimators,
+                        help=estimator_help)
     parser.add_argument('--water-level-sigma', type=float, default=3.,
                         help='Estimate max water height for each object.')
-    parser.add_argument('--known-water-threshold', type=float, default=30.,
-                        help='Threshold for extracting known water area in percent')
+    parser.add_argument('--known-water-threshold', type=optional_float, default=None,
+                        help='Threshold for extracting known water area in percent.'
+                             ' If `None`, threshold will be calculated.')
+    parser.add_argument('--minimization-metric', type=str, default='ts', choices=['fmi', 'ts'],
+                        help='Evaluation method to minimize when using the iterative estimator. '
+                             'Options include a Fowlkes-Mallows index (fmi) or a threat score (ts).')
+    parser.add_argument('--iterative-min-size', type=int, default=0,
+                        help='Minimum size of a connected waterbody in pixels for calculating flood depths with the '
+                             'iterative estimator. Waterbodies smaller than this wll be skipped.')
 
     if interface == 'hyp3':
-        parser.add_argument('--iterative-min', type=int, default=0)
-        parser.add_argument('--iterative-max', type=int, default=15)
+        parser.add_argument('--iterative-min', type=int, default=0,
+                            help='Minimum bound on the flood depths calculated using the iterative estimator.')
+        parser.add_argument('--iterative-max', type=int, default=15,
+                            help='Maximum bound on the flood depths calculated using the iterative estimator.')
     elif interface == 'main':
-        # FIXME: add a help string
-        parser.add_argument('--iterative-bounds', type=int, nargs=2, default=[0, 15])
+        parser.add_argument('--iterative-bounds', type=int, nargs=2, default=[0, 15],
+                            help='Minimum and maximum bound on the flood depths calculated using the iterative '
+                                 'estimator.')
     else:
         raise NotImplementedError(f'Unknown interface: {interface}')
 
@@ -287,7 +345,7 @@ def hyp3():
     logging.basicConfig(stream=sys.stdout, format='%(asctime)s - %(levelname)s - %(message)s', level=level)
     log.debug(' '.join(sys.argv))
 
-    if args.estimator == 'None':
+    if args.estimator is None:
         # NOTE: HyP3's current step function implementation does not have a good way of conditionally
         #       running processing steps. This allows HyP3 to always run this step but exit immediately
         #       and do nothing if flood depth maps are not requested.
@@ -315,9 +373,10 @@ def hyp3():
         out_raster=flood_map_raster, vv_raster=vv_raster, water_raster=water_map_raster, hand_raster=hand_raster,
         estimator=args.estimator, water_level_sigma=args.water_level_sigma,
         known_water_threshold=args.known_water_threshold, iterative_bounds=(args.iterative_min, args.iterative_max),
+        iterative_min_size=args.iterative_min_size, minimization_metric=args.minimization_metric,
     )
 
-    log.info(f"Flood depth map created successfully: {flood_map_raster}")
+    log.info(f'Flood depth map created successfully: {flood_map_raster}')
 
     if args.bucket:
         output_zip = make_archive(base_name=product_name, format='zip', base_dir=product_name)
@@ -334,7 +393,11 @@ def main():
     logging.basicConfig(stream=sys.stdout, format='%(asctime)s - %(levelname)s - %(message)s', level=level)
     log.debug(' '.join(sys.argv))
 
-    make_flood_map(args.out_raster, args.vv_raster, args.water_extent_map, args.hand_raster,
-                   args.estimator, args.water_level_sigma, args.known_water_threshold, tuple(args.iterative_bounds))
+    make_flood_map(
+        out_raster=args.out_raster, vv_raster=args.vv_raster, water_raster=args.water_extent_map,
+        hand_raster=args.hand_raster, estimator=args.estimator, water_level_sigma=args.water_level_sigma,
+        known_water_threshold=args.known_water_threshold, iterative_bounds=tuple(args.iterative_bounds),
+        iterative_min_size=args.iterative_min_size, minimization_metric=args.minimization_metric,
+    )
 
     log.info(f"Flood depth map created successfully: {args.out_raster}")