+++ tests/hydrosar/test_hand.py
@@ -5,8 +5,8 @@ from osgeo import gdal, ogr
 from osgeo_utils.gdalcompare import find_diff
 import numpy as np
 
-from asf_tools import hand
-
+from asf_tools import vector
+from asf_tools.hydrosar import hand
 
 HAND_BASINS = '/vsicurl/https://hyp3-testing.s3-us-west-2.amazonaws.com/' \
               'asf-tools/S1A_IW_20230228T120437_DVR_RTC30/hand/hybas_af_lev12_v1c_firstpoly.geojson'
@@ -16,8 +16,8 @@ GOLDEN_HAND = '/vsicurl/https://hyp3-testing.s3-us-west-2.amazonaws.com/' \
 gdal.UseExceptions()
 
 
-def nodata_equal_nan(GOLDEN_HAND, out_hand):
-    ds_golden = gdal.Open(str(GOLDEN_HAND))
+def nodata_equal_nan(golden_hand, out_hand):
+    ds_golden = gdal.Open(str(golden_hand))
     ds_out = gdal.Open(str(out_hand))
     nodata_golden = ds_golden.GetRasterBand(1).GetNoDataValue()
     nodata_out = ds_out.GetRasterBand(1).GetNoDataValue()
@@ -25,6 +25,7 @@ def nodata_equal_nan(GOLDEN_HAND, out_hand):
         return np.isnan(nodata_golden) and np.isnan(nodata_out)
     else:
         return False
+
 
 @pytest.mark.integration
 def test_make_copernicus_hand(tmp_path):
@@ -35,10 +36,7 @@ def test_make_copernicus_hand(tmp_path):
     assert out_hand.exists()
 
     diffs = find_diff(str(GOLDEN_HAND), str(out_hand))
-    if nodata_equal_nan(str(GOLDEN_HAND), str(out_hand)):
-        assert diffs == 1
-    else:
-        assert diffs == 0
+    assert diffs == 0
 
 
 def test_prepare_hand_vrt_no_coverage():
@@ -98,3 +96,25 @@ def test_prepare_hand_vrt_antimeridian():
 
     with pytest.raises(ValueError):
         hand.prepare_hand_vrt('foo', geometry)
+
+
+def test_intersects_hand_feature():
+    features = vector.get_features(hand.prepare.HAND_GEOJSON)
+
+    geojson = {
+        'type': 'Point',
+        'coordinates': [169, -45],
+    }
+    geometry = ogr.CreateGeometryFromJson(json.dumps(geojson))
+    assert vector.get_property_values_for_intersecting_features(geometry, features)
+
+    geojson = {
+        'type': 'Point',
+        'coordinates': [0, 0],
+    }
+    geometry = ogr.CreateGeometryFromJson(json.dumps(geojson))
+    assert not vector.get_property_values_for_intersecting_features(geometry, features)
+
+
+def test_get_features():
+    assert len(vector.get_features(hand.prepare.HAND_GEOJSON)) == 26450